Kludge for ruby 1.9 compatibility
The unit tests pass for both ruby 1.8 and 1.9, but using #ord and #chr
is neither clean nor efficient.  Aim to move away from Strings and to
byte Arrays for a cleaner implementation. This will require either an
API change or some magic (if is_a? String...) in the constructors/setters.