Remove monkey-patch for Ruby 1.8.
The monkey-patch to define String#force_encoding for Ruby 1.8 resulted
in an incompatibility with the json gem, which tries to access the
Encoding class if force_encoding is defined.  Remove this monkey-patch
and use a helper function, to_binary, that converts the string to
ASCII-8BIT encoding if encodings are available.  The byteslice function,
which was not used, has also been removed.